config updates in background
