+++ mediation.demo/src/main/java/net/pubnative/mediation/demo/activities/VideoAdActivity.java
@@ -16,7 +16,7 @@ public class VideoAdActivity extends StandardAdUnitActivity implements Pubnative
         mLoaderContainer.setVisibility(View.VISIBLE);
         PubnativeNetworkVideo video = new PubnativeNetworkVideo();
         video.setListener(this);
-        video.load(this, Settings.getAppToken(this), mPlacementSpinner.getSelectedItem().toString());
+        video.load(this, mPlacementSpinner.getSelectedItem().toString());
     }
 
     //==============================================================================================
