+++ mediation/src/main/java/net/pubnative/mediation/request/PubnativeNetworkBanner.java
@@ -10,8 +10,11 @@ import net.pubnative.mediation.adapter.PubnativeNetworkHub;
 import net.pubnative.mediation.adapter.network.PubnativeNetworkBannerAdapter;
 import net.pubnative.mediation.config.model.PubnativeNetworkModel;
 import net.pubnative.mediation.exceptions.PubnativeException;
+import net.pubnative.mediation.utils.PubnativeConfigUtils;
 
 import java.util.Map;
+
+import static net.pubnative.mediation.config.PubnativeConfigService.sIsConfigDownloading;
 
 public class PubnativeNetworkBanner extends PubnativeNetworkWaterfall
         implements PubnativeNetworkBannerAdapter.AdListener,
@@ -29,7 +32,6 @@ public class PubnativeNetworkBanner extends PubnativeNetworkWaterfall
 
         /**
          * Called whenever the banner finished loading an ad
-         * w
          *
          * @param banner banner that finished the initialize
          */
@@ -81,7 +83,7 @@ public class PubnativeNetworkBanner extends PubnativeNetworkWaterfall
     /**
      * Loads the interstitial ads before being shown
      */
-    public synchronized void load(Context context, String appToken, String placement) {
+    public synchronized void load(Context context, String placement) {
 
         Log.v(TAG, "initialize");
         if (mHandler == null) {
@@ -89,15 +91,16 @@ public class PubnativeNetworkBanner extends PubnativeNetworkWaterfall
         }
         if (mListener == null) {
             Log.w(TAG, "initialize - Warning: listener was not set, have you configured one using setListener()?");
-        }
-        if (context == null || TextUtils.isEmpty(appToken) || TextUtils.isEmpty(placement)) {
+        } else if (context == null || TextUtils.isEmpty(placement)){
             invokeLoadFail(PubnativeException.BANNER_PARAMETERS_INVALID);
+        } else if(TextUtils.isEmpty(PubnativeConfigUtils.getStoredAppToken(context)) && sIsConfigDownloading) {
+            Log.w(TAG, "initialize - Warning: config not downloaded yet, have you downloaded one using init()?");
         } else if (mIsLoading) {
             invokeLoadFail(PubnativeException.BANNER_LOADING);
         } else if (mIsShown) {
             invokeLoadFail(PubnativeException.BANNER_SHOWN);
         } else {
-            initialize(context, appToken, placement);
+            initialize(context, PubnativeConfigUtils.getStoredAppToken(context), placement);
         }
     }
