+++ mediation/src/main/java/net/pubnative/mediation/request/PubnativeNetworkFeedBanner.java
@@ -34,8 +34,11 @@ import net.pubnative.mediation.adapter.PubnativeNetworkHub;
 import net.pubnative.mediation.adapter.network.PubnativeNetworkFeedBannerAdapter;
 import net.pubnative.mediation.config.model.PubnativeNetworkModel;
 import net.pubnative.mediation.exceptions.PubnativeException;
+import net.pubnative.mediation.utils.PubnativeConfigUtils;
 
 import java.util.Map;
+
+import static net.pubnative.mediation.config.PubnativeConfigService.sIsConfigDownloading;
 
 public class PubnativeNetworkFeedBanner extends PubnativeNetworkWaterfall
         implements PubnativeNetworkFeedBannerAdapter.AdListener,
@@ -119,10 +122,9 @@ public class PubnativeNetworkFeedBanner extends PubnativeNetworkWaterfall
     /**
      * Loads the feedBanner ads before being shown
      * @param context valid context
-     * @param appToken valid app token string
      * @param placement valid placement string
      */
-    public synchronized void load(Context context, String appToken, String placement) {
+    public synchronized void load(Context context, String placement) {
 
         Log.v(TAG, "initialize");
         if (mHandler == null) {
@@ -130,18 +132,17 @@ public class PubnativeNetworkFeedBanner extends PubnativeNetworkWaterfall
         }
         if (mListener == null) {
             Log.e(TAG, "initialize - Error: listener was not set, have you configured one using setListener()?");
-        }
-        if (context == null ||
-            TextUtils.isEmpty(appToken) ||
-            TextUtils.isEmpty(placement)) {
+        } else if (context == null || TextUtils.isEmpty(placement)){
             invokeLoadFail(PubnativeException.FEED_BANNER_PARAMETERS_INVALID);
+        } else if(TextUtils.isEmpty(PubnativeConfigUtils.getStoredAppToken(context)) && sIsConfigDownloading) {
+            Log.w(TAG, "initialize - Warning: config not downloaded yet, have you downloaded one using init()?");
         } else if (mIsLoading) {
             invokeLoadFail(PubnativeException.FEED_BANNER_LOADING);
         } else if (mIsShown) {
             invokeLoadFail(PubnativeException.FEED_BANNER_SHOWN);
         } else {
             mIsLoading = true;
-            initialize(context, appToken, placement);
+            initialize(context, PubnativeConfigUtils.getStoredAppToken(context), placement);
         }
     }
