+++ mediation/src/main/java/net/pubnative/mediation/request/PubnativeNetworkVideo.java
@@ -31,11 +31,13 @@ import android.util.Log;
 
 import net.pubnative.mediation.adapter.PubnativeNetworkHub;
 import net.pubnative.mediation.adapter.network.PubnativeNetworkVideoAdapter;
-import net.pubnative.mediation.adapter.network.PubnativeNetworkVideoAdapter;
 import net.pubnative.mediation.config.model.PubnativeNetworkModel;
 import net.pubnative.mediation.exceptions.PubnativeException;
+import net.pubnative.mediation.utils.PubnativeConfigUtils;
 
 import java.util.Map;
+
+import static net.pubnative.mediation.config.PubnativeConfigService.sIsConfigDownloading;
 
 public class PubnativeNetworkVideo extends PubnativeNetworkWaterfall
         implements PubnativeNetworkVideoAdapter.LoadListener,
@@ -130,25 +132,24 @@ public class PubnativeNetworkVideo extends PubnativeNetworkWaterfall
     /**
      * Loads the video ads before being shown
      * @param context valid Context
-     * @param appToken valid app token string
      * @param placement valid placement string
      */
-    public synchronized void load(Context context, String appToken, String placement) {
+    public synchronized void load(Context context, String placement) {
 
         Log.v(TAG, "initialize");
         if (mListener == null) {
             Log.e(TAG, "initialize - Error: listener was not set, have you configured one using setListener()?");
-        } else if (context == null ||
-                   TextUtils.isEmpty(appToken) ||
-                   TextUtils.isEmpty(placement)) {
+        } else if (context == null || TextUtils.isEmpty(placement)){
             invokeLoadFail(PubnativeException.VIDEO_PARAMETERS_INVALID);
+        } else if(TextUtils.isEmpty(PubnativeConfigUtils.getStoredAppToken(context)) && sIsConfigDownloading) {
+            Log.w(TAG, "initialize - Warning: config not downloaded yet, have you downloaded one using init()?");
         } else if (mIsLoading) {
             invokeLoadFail(PubnativeException.VIDEO_LOADING);
         } else if (mIsShown) {
             invokeLoadFail(PubnativeException.VIDEO_SHOWN);
         } else {
             mHandler = new Handler(Looper.getMainLooper());
-            initialize(context, appToken, placement);
+            initialize(context, PubnativeConfigUtils.getStoredAppToken(context), placement);
         }
     }
